+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.90.23
+- Added Incident Reporting
+- Added Cypress tests
+
 # 0.90.22
 - Added range input types for integer form fields
 - Modified Sentry error logging details